Extract Smart Race Solver helpers into src/lib/solver constants and scoring modules
